refactor: member/dto 패키지 구조 개선(response/request로 분리)
