+++ backend/src/main/java/org/example/be/member/dto/request/MemberPolicyUpdateReqBody.java
@@ -1,4 +1,4 @@
-package org.example.be.member.dto;
+package org.example.be.member.dto.request;
 
 import jakarta.validation.constraints.NotNull;
